+++ src/main/java/tqs/sparkflow/userservice/UserServiceApplication.java
@@ -1,4 +1,4 @@
-package tqs.sparkflow.user_service;
+package tqs.sparkflow.userservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
