+++ src/main/java/tqs/sparkflow/userservice/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package tqs.sparkflow.user_service.config;
+package tqs.sparkflow.userservice.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
