+++ src/test/java/tqs/sparkflow/userservice/config/ConfigBeansTest.java
@@ -1,4 +1,4 @@
-package tqs.sparkflow.user_service.config;
+package tqs.sparkflow.userservice.config;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -8,6 +8,10 @@ import org.springframework.test.context.DynamicPropertySource;
 import org.testcontainers.containers.MongoDBContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
+
+import tqs.sparkflow.userservice.config.OpenApiConfig;
+import tqs.sparkflow.userservice.config.SecurityConfig;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
