+++ src/test/java/tqs/sparkflow/userservice/config/TestSecurityConfig.java
@@ -1,4 +1,4 @@
-package tqs.sparkflow.user_service.config;
+package tqs.sparkflow.userservice.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
