+++ src/test/java/tqs/sparkflow/userservice/service/UserServiceTest.java
@@ -1,4 +1,4 @@
-package tqs.sparkflow.user_service.service;
+package tqs.sparkflow.userservice.service;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
@@ -15,10 +15,11 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import tqs.sparkflow.user_service.exception.DuplicateEmailException;
-import tqs.sparkflow.user_service.exception.ResourceNotFoundException;
-import tqs.sparkflow.user_service.model.User;
-import tqs.sparkflow.user_service.repository.UserRepository;
+import tqs.sparkflow.userservice.exception.DuplicateEmailException;
+import tqs.sparkflow.userservice.exception.ResourceNotFoundException;
+import tqs.sparkflow.userservice.model.User;
+import tqs.sparkflow.userservice.repository.UserRepository;
+import tqs.sparkflow.userservice.service.UserService;
 
 @ExtendWith(MockitoExtension.class)
 class UserServiceTest {
