+++ src/test/resources/application.properties
@@ -1,6 +1,11 @@
 # Test Configuration
-spring.data.mongodb.auto-index-creation=true
+spring.data.mongodb.uri=mongodb://localhost:27017/test
 spring.data.mongodb.database=test
+spring.data.mongodb.auto-index-creation=true
+
+# Disable security for tests
+spring.security.user.name=test
+spring.security.user.password=test
 
 # Server Configuration
 server.port=0
